Change Fabulous license to Apache 2.0
This change simplifies things. The bundled Grapefruit library is already
Apache 2.0, so it makes little sense to be both MIT and Apache 2.0.

Part of the goal here is also to adopt Google's best practices for open
source licensing, which is basically using Apache 2.0, having copyright
headers, and using an AUTHORS / CONTRIBUTORS dichotomy. As a result,
past contributors have been added to these files so they get credit for
their work. However Fabulous will not require a CLA going forward.

I apologize to users with GPLv2 project. But you were already in
violation by transitively linking the Apache 2.0 library Grapefruit,
which is probably a violation. So you're about is incompliant now as you
were before. Consider upgrading to GPLv3, which makes an exception for
Apache 2.0.